Class "Rcpp_corrData"

This C++ class could be used to build a corrData object by invoking new(corrData,...) function.

References

See the documentation for Rcpp modules for more details of how this class was built. vignette("Rcpp-modules", package = "Rcpp")

Note

This is only frame for building C++ object which could be used to implement certain methods. Check the vignette for more details of implementing methods.

Vigniette: https://CRAN.R-project.org/package=miceFast

Methods

initialize(...):

~~

finalize():

~~

fill(...):

generating data

Examples

#showClass("Rcpp_corrData")
show(corrData)
#> C++ class 'corrData' <0x600003010600>
#> Constructors:
#>     corrData(int, arma::Col<double>, arma::Mat<double>)
#>     corrData(int, int, arma::Col<double>, arma::Mat<double>)
#> 
#> Fields: No public fields exposed by this class
#> 
#> Methods: 
#>      arma::Mat<double> fill(std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char>>)  
#>