Class "Rcpp_miceFast"

This C++ class could be used to build a miceFast objects by invoking new(miceFast) function.

Extends

Class "C++Object", directly.

All reference classes extend and inherit methods from "envRefClass".

References

See the documentation for RcppArmadillo and Rcpp for more details of how this class was built.

Note

This is only frame for building C++ object which could be used to implement certain methods. Check the vignette for more details of implementing these methods.

Vigniette: https://CRAN.R-project.org/package=miceFast

Methods

set_data(...):

providing data by a reference - a numeric matrix

set_g(...):

providing a grouping variable by a reference - a numeric vector - positive values

set_w(...):

providing a weightinh variable by a reference - a numeric vector - positive values

get_data(...):

retrieving the data

get_w(...):

retrieving the weighting variable

get_g(...):

retireiving the grouping variable

get_index(...):

getting the index

impute(...):

impute data under characterstics from the object like a optional grouping or weighting variable

impute_N(...):

multiple imputations - impute data under characterstics from the object like a optional grouping or weighting variable

update_var(...):

permanently update the variable at the object and data. Use it only if you are sure about model parameters

get_models(...):

get possible quantitative models for a certain type of dependent variable

get_model(...):

get a recommended quantitative model for a certain type of dependent variable

which_updated(...):

which variables at the object was modified by update_var

sort_byg(...):

sort data by the grouping variable

is_sorted_byg(...):

check if data is sorted by the grouping variable

vifs(...):

Variance inflation factors (VIF) - helps to check when the predictor variables are not linearly related

initialize(...):

...

finalize():

...

Examples

#showClass("Rcpp_miceFast")
show(miceFast)

#> C++ class 'miceFast' <0x7f928dc4b670>
#> Constructors:
#>     miceFast()
#> 
#> Fields: No public fields exposed by this class
#> 
#> Methods: 
#>      arma::Mat<double> get_data()  
#>            
#>      arma::Col<double> get_g()  
#>            
#>      arma::Col<unsigned int> get_index()  
#>            
#>      std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char> > get_model(int)  
#>            
#>      std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char> > get_models(int)  
#>            
#>      double get_ridge()  
#>            
#>      arma::Col<double> get_w()  
#>            
#>      Rcpp::List impute(std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char> >, int, arma::Col<unsigned int>)  
#>            
#>      Rcpp::List impute_N(std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char> >, int, arma::Col<unsigned int>, int)  
#>            
#>      bool is_sorted_byg()  
#>            
#>      void set_data(arma::Mat<double>)  
#>            
#>      void set_g(arma::Col<double>)  
#>            
#>      void set_ridge(double)  
#>            
#>      void set_w(arma::Col<double>)  
#>            
#>      void sort_byg()  
#>            
#>      void update_var(int, arma::Col<double>)  
#>            
#>      arma::Col<double> vifs(int, arma::Col<unsigned int>)  
#>            
#>      std::__1::vector<int, std::__1::allocator<int> > which_updated()  
#>            

new(miceFast)

#> C++ object <0x7f928e7aae00> of class 'miceFast' <0x7f928dc4b670>