fill_NA function for the imputations purpose.

Regular imputations to fill the missing data. Non missing independent variables are used to approximate a missing observations for a dependent variable. Quantitative models were built under Rcpp packages and the C++ library Armadillo.

auto_fill_NA(x)

Arguments

x
a numeric matrix or data.frame/data.table (factor/character/numeric/logical) - variables

Value

load imputations in a numeric/character/factor (similar to the input type) vector format

Note

There is assumed that users add the intercept by their own. The miceFast module provides the most efficient environment, the second recommended option is to use data.table and the numeric matrix data type. The lda model is assessed only if there are more than 15 complete observations and for the lms models if number of independent variables is smaller than number of observations.

See also

Examples

if (FALSE) {
library(miceFast)
data(air_miss)
auto_fill_NA(air_miss)
}