naive_fill_NA function for the imputations purpose.

Regular imputations to fill the missing data. Non missing independent variables are used to approximate a missing observations for a dependent variable. For numeric columns with any missing data a simple bayesian mean will be used. Next all numeric variables will be utilized to impute character/factoer variables by Linear Discriminant Analysis.

naive_fill_NA(x)

# S3 method for data.frame
naive_fill_NA(x)

# S3 method for matrix
naive_fill_NA(x)

Arguments

x
a numeric matrix or data.frame/data.table (factor/character/numeric/logical) - variables

Value

load imputations in a similar to the input type

Methods (by class)

data.frame: S3 method for data.frame

matrix: S3 method for matrix

Note

this is a very simple and fast solution but not recommended, for more complex solutions check the vignette

See also

Examples

if (FALSE) {
library(naive_fill_NAast)
data(air_miss)
naive_fill_NA(air_miss)
}